pyproject.toml
[project]
name = "apprentice"
version = "0.4.0"
description = "Agentic Algorithm Factory for no-magic"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "google-adk>=1.0.0",
    "jinja2>=3.1,<4",
    "litellm>=1.50.0",
    "openai>=2.30.0",
    "pyyaml>=6.0.3",
    "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
apprentice = "apprentice.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apprentice"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific rules
]
ignore = [
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["apprentice"]

[tool.mypy]
strict = true
python_version = "3.12"
mypy_path = "src"
packages = ["apprentice"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "types-pyyaml>=6.0.12.20250915",
]
